Ensure preditctor column of splits is a factor
In R v4.0.0 and above, stringsAsFactors is now FALSE by default.

The getSplitImprove*() functions did not explicitly create a
factor, but later functions such as combinedGradientForest assume
that it is a factor.

Now getSplitImprove*() functions explicity cast to factor, with
levels for all predictors passed to gradientForest(), even if a
predictor is never used in a split.