Allow combinedGradientForest with dropped preds
It is possible that gradientForest never uses
a predictor in any split during fitting.

If a gradientForest object left one or
more predictors unused, then it threw an
error when you try to merge with another
gradientForest object that had a different
set of used predictors.

The predictors passed to each gradientForest for fitting
must still match. This fix just allows combinations
of gradientForest objects that use different subsets
of the predictors.