Allow combining with different predictors
This patch attempts to allow combinedGradientForest
to work on gradientForest objects that do not use
the same predictors.

It should work if the input predictors are different,
and if the final set of used predictors are different.

End users will need to take care that predictors are
given the same names between gradientForest objects, mismatched
names will be treated as unrelated predictors.