Fix all downstream functions
combinedGradientForest now accepts any combination
of predictors, both as inputs to the individual
gradientForest objects, and in the set of predictors
used by each gradientForest (that is, a predictor
is used in at least one split in the random forest).

However, many downstream functions made assumptions about
the structure of combinedGradientForest that are no
longer valid, usually that X is a large matrix containing
all the sites from all the member gradientForest objects.
These functions are now updated.